% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bidi5}
\alias{bidi5}
\alias{bidi6}
\title{Bidirected graphs on 5 and 6 vertices}
\format{
An object of class \code{list} of length 21.
}
\usage{
bidi5
}
\description{
List of all bidirected connected graphs on 5 and 6 unlabelled vertices.
}
\keyword{datasets}
