% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_decomp_ZH.R
\name{bidi_decomposition}
\alias{bidi_decomposition}
\title{Implement algorithm from Hu and Evans (2022)}
\usage{
bidi_decomposition(graph)
}
\arguments{
\item{graph}{bidirected graph of class \code{mixedgraph}}
}
\description{
Implement algorithm from Hu and Evans (2022)
}
\details{
Implements the algorithm described in Section 5 of Hu and Evans
(2022).
}
\examples{
data(bidi5)
bidi_decomposition(bidi5[[1]])
bidi_decomposition(bidi5[[1]])

}
