% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsets.R
\name{char_imset}
\alias{char_imset}
\title{Characteristic imset (for ADMGs...)}
\usage{
char_imset(x, ...)
}
\arguments{
\item{x}{a graph in ADMGs format or an imset}
}
\description{
Characteristic imset (for ADMGs...)
}
