% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andrews3.R
\name{ci2andrews}
\alias{ci2andrews}
\alias{andrews2ci}
\title{Convert between ci object and text representation}
\usage{
ci2andrews(cis, symm = FALSE)

andrews2ci(x)
}
\arguments{
\item{cis}{list of ci objects}

\item{symm}{logical: should symmetric independences be listed}

\item{x}{string of Andrew's notation}
}
\description{
Convert between ci object and text representation
}
\section{Functions}{
\itemize{
\item \code{andrews2ci}: convert text string to ci objects
}}

