% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cons_imset.R
\name{cons_imset}
\alias{cons_imset}
\title{Get the imset representing constrained sets from a conditional independence}
\usage{
cons_imset(ci, n)
}
\arguments{
\item{ci}{a conditional independence as an object of class \code{ci}}

\item{n}{total number of variables}
}
\description{
Get the imset representing constrained sets from a conditional independence
}
