% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP.R
\name{defines_mod}
\alias{defines_mod}
\title{Check if the standard imset for a graph defines the model}
\usage{
defines_mod(
  graph,
  u,
  timeout = 60L,
  trace = FALSE,
  sparse = FALSE,
  rmv_edges = FALSE,
  stop_if_fail = TRUE
)
}
\arguments{
\item{graph}{an ADMG in the form of a \code{mixedgraph} object}

\item{u}{optionally, an imset to test the constraints from the graph on
(otherwise the 'standard' imset is used)}

\item{timeout}{an integer giving the maximum run time for each linear program in seconds}

\item{trace}{logical: should details be given of tests?}

\item{sparse}{logical: use sparse matrix?}

\item{rmv_edges}{logical: only use missing edges in the graph in the constraint matrix?}

\item{stop_if_fail}{logical: should procedure stop if one independence is not
represeneted?}
}
\description{
Check if the standard imset for a graph defines the model
}
\details{
The \code{timeout} argument requires version 5.6.13.4.9000 of \code{lpSolve} to be
installed.  If the function returns \code{NA} then no independence was found
to be not in the model but at least one of the linear programs timed out.
}
