% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP.R
\name{definesMod}
\alias{definesMod}
\title{Check if the standard imset for a graph defines the model}
\usage{
definesMod(graph, timeout = 60L)
}
\arguments{
\item{graph}{an ADMG in the form of a \code{mixedgraph} object}

\item{timeout}{an integer giving the maximum run time for each linear program in seconds}
}
\description{
Check if the standard imset for a graph defines the model
}
\details{
This requires version 5.6.13.4 or higher of \code{lpSolve} to be
installed.  If the function returns \code{NA} then no independence was found
to be not in the model but at least one of the linear programs timed out.
}
