% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementary_imsets.R
\name{elemImset-deprecated}
\alias{elemImset-deprecated}
\alias{elemImset}
\alias{elemToIndep}
\title{Deprecated function for (semi-)elementary imsets}
\usage{
elemImset(A, B, C = integer(0), n = max(c(A, B, C)))

elemToIndep(imset)
}
\arguments{
\item{A, B, C}{disjoint subsets of \eqn{1,...,n}}

\item{n}{number of variables involved}
}
\description{
Deprecated function for (semi-)elementary imsets
}
\details{
Returns an imset, a vector of length \eqn{2^n}
with all but four entries zero.

Function is now deprecated in favour of \code{elem_imset}.
}
\section{Functions}{
\itemize{
\item \code{elemToIndep()}: Deprecated version of elemToIndep, use \code{sei2ci}

}}
