% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsets.R
\name{elemImset}
\alias{elemImset}
\alias{elemToIndep}
\title{Get a (semi-)elementary imset}
\usage{
elemImset(A, B, C = integer(0), n = max(c(A, B, C)))

elemToIndep(imset)
}
\arguments{
\item{A, B, C}{disjoint subsets of 1,..,n}

\item{n}{number of variables involved}
}
\description{
Returns the semi-elementary imset \eqn{u_{\langle A, B | C \rangle}}{u<A, B | C>}
as defined in Studeny (2006).
}
\details{
Returns an imset, a vector of length \eqn{2^n}
with all but four entries zero.
}
\section{Functions}{
\itemize{
\item \code{elemToIndep}: Construct conditional independence from elementary imset
}}

