% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsets.R
\name{envelopes}
\alias{envelopes}
\alias{upperEnvelope}
\alias{lowerEnvelope}
\title{Get maximal positive/negative sets from imset}
\usage{
upperEnvelope(x)

lowerEnvelope(x)
}
\arguments{
\item{x}{imset}
}
\description{
Intended for a structural imset, obtain the maximal positive and negative
sets in an imset.
}
\section{Functions}{
\itemize{
\item \code{upperEnvelope()}: Get maximal positive sets from imset

\item \code{lowerEnvelope()}: Get maximal negative sets from imset

}}
