% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fail56}
\alias{fail56}
\title{MAGs that do not defined the model over 5 and 6 vertices}
\format{
An object of class \code{list} of length 56.
}
\usage{
fail56
}
\description{
MAGs that do not defined the model over 5 and 6 vertices
}
\keyword{datasets}
