% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr1a}
\alias{gr1a}
\alias{gr1b}
\title{Pair of bidirected graphs}
\format{
An object of class \code{mixedgraph} of length 3.
}
\usage{
gr1a
}
\description{
Two graphs, \code{gr1a} and \code{gr1b}, that differ by one edge, but one
(\code{gr1a}) is perfectly Markovian with respect to its standard imset,
and the other (\code{gr1b}) is not.
}
\keyword{datasets}
