% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{imset_arith}
\alias{imset_arith}
\alias{+.imset}
\alias{-.imset}
\title{Overload arithmetic for imsets}
\usage{
\method{+}{imset}(e1, e2)

\method{-}{imset}(e1, e2)
}
\arguments{
\item{e1, e2}{two imsets}
}
\description{
Overload arithmetic for imsets
}
\details{
These functions are important, because they prevent accidentally
adding imsets with a different number of variables to one another, and the
recycling in R giving the wrong answer.
}
\section{Functions}{
\itemize{
\item \code{+.imset}: Add imsets

\item \code{-.imset}: Subtract imsets
}}

