% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{imset_control}
\alias{imset_control}
\alias{as.imset}
\alias{is.imset}
\title{Test and enforce that vector is an imset}
\usage{
as.imset(x)

is.imset(x)
}
\arguments{
\item{x}{vector test or make imset from}
}
\description{
Test and enforce that vector is an imset
}
\details{
\code{as.imset} forces a vector to become an imset;
the length of the vector \code{x} must be a power of 2 length.
\code{is.imset} just checks that \code{imset} is in
\code{class(x)}.
}
\section{Functions}{
\itemize{
\item \code{as.imset()}: force object to be an imset

\item \code{is.imset()}: test if object is an imset

}}
