% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ints.R
\name{intSubset}
\alias{intSubset}
\title{Compare subsets by integer position in vector}
\usage{
intSubset(x, y)
}
\arguments{
\item{x, y}{values to compare}
}
\description{
Compare subsets by integer position in vector
}
