% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP.R
\name{is_combinatorial}
\alias{is_combinatorial}
\alias{is_structural}
\alias{imset_degree}
\title{Check if the an imset is structural or combinatorial}
\usage{
is_combinatorial(imset, timeout = 60L, sparse = FALSE)

is_structural(imset, timeout = 60L, sparse = FALSE)

imset_degree(imset, timeout = 60L)
}
\arguments{
\item{imset}{an imset to be tested}

\item{timeout}{an integer giving the maximum run time for each linear program in seconds}

\item{sparse}{should sparse matrices be used?}
}
\description{
Uses linear programming to check whether imsets can be decomposed as
elementary imsets.
}
\details{
For the \code{timeout} operator to work we require version 5.6.13.4.9000
of \code{lpSolve} to be installed.
}
\section{Functions}{
\itemize{
\item \code{is_structural()}: test if imset is structural

\item \code{imset_degree()}: for a structural imset, get its degree

}}
