% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_imset.R
\name{isCombinatorial}
\alias{isCombinatorial}
\title{Recursive function to check if an imset if combinatorial}
\usage{
isCombinatorial(imset)
}
\arguments{
\item{imset}{imset to be checked}
}
\description{
Recursive function to check if an imset if combinatorial
}
\details{
This function recursively tries to remove
}
