% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_imsets_S3.R
\name{make_imset}
\alias{make_imset}
\title{Construct sparse imset}
\usage{
make_imset(x, pos, n, vnames)
}
\arguments{
\item{x}{entries}

\item{pos}{indices or list of subsets}

\item{n}{number of variables}

\item{vnames}{names for variables}
}
\description{
Construct sparse imset
}
