% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andrews3.R
\name{NIE}
\alias{NIE}
\title{Implement Andrews' Algorithm 6}
\usage{
NIE(graph, topOrd)
}
\arguments{
\item{graph}{an ADMG of class \code{mixedgraph}}

\item{topOrd}{an optional topological ordering}
}
\description{
Implement Andrews' Algorithm 6
}
\details{
Implements Algorithm 6 from Andrews (2022).
}
