% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsets.R
\name{+.imset}
\alias{+.imset}
\title{Overload addition for imsets}
\usage{
\method{+}{imset}(e1, e2)
}
\arguments{
\item{e1, e2}{two imsets}
}
\description{
Overload addition for imsets
}
\details{
This function is important, because it prevents accidentally
adding imsets with a different number of variables to one another, and the
recycling in R giving the wrong answer.
}
