% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ints.R
\name{position_subset}
\alias{position_subset}
\alias{posToSubset}
\alias{subsetToPos}
\title{Move between subsets and integer position in vector}
\usage{
posToSubset(x, simplify = FALSE)

subsetToPos(y)
}
\arguments{
\item{x}{vector of integers to compare}

\item{simplify}{if \code{x} has length one, just return a vector}

\item{y}{list of subsets to consider}
}
\description{
Move between subsets and integer position in vector
}
\section{Functions}{
\itemize{
\item \code{posToSubset()}: Obtain subset from position

\item \code{subsetToPos()}: Integer position from subset

}}
