% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ints.R
\name{posToSubset}
\alias{posToSubset}
\title{Compute subsets by integer position in vector}
\usage{
posToSubset(x, simplify = FALSE)
}
\arguments{
\item{x}{vector of integers to compare}

\item{simplify}{if \code{x} has length one, just return a vector}
}
\description{
Compute subsets by integer position in vector
}
