% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andrews.R
\name{set_entropy}
\alias{set_entropy}
\title{Compute Andrews' summation}
\usage{
set_entropy(graph, p)
}
\arguments{
\item{graph}{a \code{mixedgraph} object of summary graph form}

\item{p}{a distribution or list or tables object of distributions}
}
\value{
A numeric vector with the same length as the number of distributions
supplied.
}
\description{
Compute the right-hand side of Andrews' entropy formula.
}
\details{
This function uses the parameterizing sets to compute the entropy
with respect to each distribution in \code{p}.  If the distribution is in
the model then this will be equal to the entropy.
}
