% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsets.R
\name{setSign}
\alias{setSign}
\title{Assign signs to head-tail sets}
\usage{
setSign(h, t)
}
\arguments{
\item{h, t}{head and tail sets

Assign -1 to sets of even size, 1 to sets of odd size.}
}
\description{
Assign signs to head-tail sets
}
