% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{setSize}
\alias{setSize}
\title{Vector of sizes for power set}
\usage{
setSize(n, m = n)
}
\arguments{
\item{n}{size of set}

\item{m}{maximum size to consider}
}
\description{
Vector of sizes for power set
}
\details{
Uses Kronecker sum to obtain sizes of subsets in a conventionally
ordered power set.
}
