% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsets.R
\name{standard_imset}
\alias{standard_imset}
\title{Standard imset (for DAGs only)}
\usage{
standard_imset(x, ...)
}
\arguments{
\item{x}{a graph in ADMGs format or an imset}
}
\description{
Standard imset (for DAGs only)
}
