% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{tailPowSet}
\alias{tailPowSet}
\title{Sets of form HuA for A subset of T}
\usage{
tailPowSet(h, t)
}
\arguments{
\item{h, t}{head and tail sets}
}
\description{
Sets of form HuA for A subset of T
}
\details{
Returns a list of sets of the form
\eqn{H \cup A} where \eqn{A \subseteq T}.
}
