% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP.R
\name{test_indep}
\alias{test_indep}
\title{Test if an independence is represented in an imset}
\usage{
test_indep(imset, ci, timeout = 60L)
}
\arguments{
\item{imset}{an imset}

\item{ci}{a conditional independence}

\item{timeout}{a timeout for the linear program solver in seconds}
}
\description{
Test if an independence is represented in an imset
}
\details{
Runs linear program suggested in Lindner (2012), Section 6.3.2. The
timeout variable defaults to 60 seconds: using 0 means there is no limit.
}
\references{
Lindner (2012). \emph{Discrete optimisation in machine learning-learning of Bayesian network structures and conditional independence implication}, PhD Thesis, TUM.
}
