% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsets.R
\name{char_imset.imset}
\alias{char_imset.imset}
\title{Get characteristic imset from standard}
\usage{
char_imset.imset(x)
}
\arguments{
\item{x}{a standard imset

maybe make this generic to work with graphs and
other imsets?}
}
\description{
Get characteristic imset from standard
}
