% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP.R
\name{is_combinatorial}
\alias{is_combinatorial}
\alias{is_structural}
\title{Check if the an imset is structural or combinatorial}
\usage{
is_combinatorial(imset, timeout = 60L)

is_structural(imset, timeout = 60L)
}
\arguments{
\item{imset}{an imset to be tested}

\item{timeout}{an integer giving the maximum run time for each linear program in seconds}
}
\description{
Uses linear programming to check whether imsets can be decomposed as
elementary imsets.
}
\details{
This requires version 5.6.13.4 or higher of \code{lpSolve} to be
installed.  If the function returns \code{NA} then no independence was found
to be not in the model but at least one of the linear programs timed out.
}
\section{Functions}{
\itemize{
\item \code{is_structural}: test if imset is structural
}}

\seealso{
\link{\code{defines_mod}}
}
