% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ints.R
\name{wh_entries}
\alias{wh_entries}
\title{Get entries in vector of subsets}
\usage{
wh_entries(x)
}
\arguments{
\item{x}{list of positive integer subsets

Given a subset of 1,2,3,... returns the location
where set would appear in lexicographic order.}
}
\description{
Gives location where set would appear in imset under
lexicographic ordering.
}
\examples{
imset:::wh_entries(list(1, 3, 1:3))
imset:::wh_entries(powerSet(1:3))
imset:::wh_entries(1:3)

}
