% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsets.R
\name{zero_imset}
\alias{zero_imset}
\alias{identifier_imset}
\title{Zero imset}
\usage{
zero_imset(n, sparse = FALSE, vnames)

identifier_imset(A, n, sparse = FALSE, vnames)
}
\arguments{
\item{n}{number of variables}

\item{sparse}{logical: should sparse imsets be used?}

\item{A}{set to identify}
}
\description{
Zero imset
}
\section{Functions}{
\itemize{
\item \code{identifier_imset()}: identify a single set

}}
